Implement controller method
To determine which controllers will be supported by the api we implement
a method which returns all controller names.
To test the module we also add some simple tests that can be run against
an existing phpIPAM installation.
Here you only need to create a `server.yml` in `tests/vars`. Simply copy
the `server.yml.exampl` and fill out with your server data.
To secure your secrets we add `server.yml` to `.gitignore`