Record tests for replay in CI
All tests will now be recorded to let it run in CI without interaction
with a real phpIPAM instance.
For that we import `vcr` in all and `pytest` in some test cases. Each
test get a `vcr.use_casssette` decorator to save the cassette as
expected.
In `conftest` we define some methods to sanitize the recordings and one
method to create the cassette name.