Fix #38 - http vs. https connections
As the try to connect to an unsecure api endpoint with user token leads
to an exception we need to adapt our exception handling to print out the
full and correct information.
As `https` connections are industry standard we don't invest much effort
in implementation of the `encryption` feature.
If anybody needs this feature feel free to contribute or file a feature
request where interested can vote on.