Security/fix requests version (#59)
* Increase requests minimum version

As of security warnings in older `requests` version we decided to
increase the minimum version of that libery in all requirements.

* Set explicitly ubuntu to 20.04 for test

As in current `ubuntu-latest` not all python version are available we
switch explicitly back to `ubunt-20.04`.
We also remove `python 3.6` from supported python distributions because
it does not safely provide needed requests version.